Fix issue with code generation on Windows
Apparently older versions of the "openapi-generator-maven-plugin" library could not handle "\" so paths on Windows failed.

NOTE: It still does not play nice with spaces in the path! So if the path to the base folder where you keep your efm-integrasjonspunkt repo is something like "C:\my job projects for work" then the efm-integrasjonspunkt project will not build!